Update trailData-slice with newTrailPreview  state
